+++ coralogix/resource_coralogix_group.go
@@ -20,8 +20,9 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"terraform-provider-coralogix/coralogix/clientset"
+
+	"github.com/hashicorp/terraform-plugin-framework/attr"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
@@ -90,6 +91,10 @@ func (r *GroupResource) Schema(_ context.Context, _ resource.SchemaRequest, resp
 			"role": schema.StringAttribute{
 				Required: true,
 			},
+			"scope_id": schema.StringAttribute{
+				Optional:            true,
+				MarkdownDescription: "Scope attached to the group.",
+			},
 		},
 		MarkdownDescription: "Coralogix group.",
 	}
@@ -123,10 +128,10 @@ func (r *GroupResource) Create(ctx context.Context, req resource.CreateRequest, 
 		)
 		return
 	}
-	groupStr, _ = json.Marshal(createResp)
-	log.Printf("[INFO] Submitted new group: %s", groupStr)
-
-	state, diags := flattenSCIMGroup(createResp)
+	getResp, err := r.client.GetGroup(ctx, createResp.ID)
+	groupStr, _ = json.Marshal(getResp)
+	log.Printf("[INFO] Getting group: %s", groupStr)
+	state, diags := flattenSCIMGroup(getResp)
 	if diags.HasError() {
 		resp.Diagnostics.Append(diags...)
 		return
@@ -143,11 +148,17 @@ func flattenSCIMGroup(group *clientset.SCIMGroup) (*GroupResourceModel, diag.Dia
 		return nil, diags
 	}
 
+	scopeId := types.StringNull()
+	if group.ScopeID != "" {
+		scopeId = types.StringValue(group.ScopeID)
+	}
+
 	return &GroupResourceModel{
 		ID:          types.StringValue(group.ID),
 		DisplayName: types.StringValue(group.DisplayName),
 		Members:     members,
 		Role:        types.StringValue(group.Role),
+		ScopeID:     scopeId,
 	}, nil
 }
 
@@ -297,6 +308,7 @@ type GroupResourceModel struct {
 	DisplayName types.String `tfsdk:"display_name"`
 	Members     types.Set    `tfsdk:"members"` // Set of strings
 	Role        types.String `tfsdk:"role"`
+	ScopeID     types.String `tfsdk:"scope_id"`
 }
 
 func extractGroup(ctx context.Context, plan *GroupResourceModel) (*clientset.SCIMGroup, diag.Diagnostics) {
@@ -309,6 +321,7 @@ func extractGroup(ctx context.Context, plan *GroupResourceModel) (*clientset.SCI
 		DisplayName: plan.DisplayName.ValueString(),
 		Members:     members,
 		Role:        plan.Role.ValueString(),
+		ScopeID:     plan.ScopeID.ValueString(),
 	}, nil
 }