+++ CHANGELOG.md
@@ -609,8 +609,17 @@ New Features:
 * added support for more than one `section`.
 * added support for `query` option in `multi_select` variables.
 
-
 ## Release 1.16.4
 Bug fixing:
 #### resource/coralogix_dashboard
 * changing `pie_chart` and `horizontal_bar_chart` `query.logs.group_names` to Optional.
+
+## Release 1.17
+
+New Features:
+#### resource/coralogix_scope
+* added Scope support
+
+#### resource/coralogix_group
+* added support for associated scopes
+